Add BOM test for JSON and CSS module scripts
Add tests to validate that CSS and JSON module source is always decoded
as UTF-8 even if a UTF-16 BOM is present.

For JSON module scripts, this results in parse errors for UTF-16
sources. The CSS parser is more forgiving and doesn't throw an error,
but instead fills in the invalid chars in the UTF-8 decoded source
with U+FFFD REPLACEMENT CHARACTER.

Bug: 992499,1132413
Change-Id: I91e2f79bf3cc44fd47d9b80e49b1285d452bb9ff