Cache Storage: Fix null deref with explicit undefined passed
Given the IDL definition of two methods with args of the form
(optional request, optional options), the bindings code expects two
overloads in the implementation. The bindings code maps calls to the
overloads differently if an explicit undefined is passed for request
with no options vs. an explicit undefined is passed with an options
arg, and the code wasn't expecting
the request to be null.

Bug: 1232955
Change-Id: I2ea29151dfaf4e764f221200ef4acb4784334a34